Create Pcut_function.py
Pcut function computes the most frequent labels in a subset and assigns the final labels by considering the Label cardinality (LC) of the dataset.
For example, let consider a dataset with LC=4, first, Pcut computes the probability of an example to have each label in a subset of learning and sort them in a descending manner, and finally, it keeps four active labels with high probabilities (=LC).
This function has been used in many fields, you can find an example of use in a Personalized Decision Trees in our repository: LaCova-Python, where this function is used by the main authors to annotate the leaves. It has been used also for text categorization, as in [1]
[1] Yang, Y. (2001, September). A study of thresholding strategies for text categorization. In Proceedings of the 24th annual international ACM SIGIR conference on Research and development in information retrieval (pp. 137-145).